requirements.txt
# Core PyTorch ecosystem
torch>=1.9.0
torchvision>=0.10.0
torchsummary>=1.5.1

# TensorBoard and logging
tensorboardX>=2.2
tensorboard>=2.8.0

# Image processing and computer vision
torchmetrics[image]>=0.9.0
pytorch-msssim>=0.2.1
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Progress bars and utilities
tqdm>=4.62.0

# Configuration and data handling
PyYAML>=5.4.0

# Visualization
matplotlib>=3.3.0

# Additional dependencies for advanced features
pathlib2>=2.3.0; python_version < "3.4"